// src/Styleguide/types.ts
export enum Size {
  Zero = 'zero',
  Tiny = 'tiny',
  Small = 'small',
  Medium = 'medium',
  Large = 'large',
  ExtraLarge = 'xLarge',
}

export enum Variant {
  Primary = 'primary',
  Secondary = 'secondary',
  Tertiary = 'tertiary'
}

export enum Status {
  SoldOut = 'soldOut',
  Discount = 'discount',
  Default = 'default',
  Success = 'success',
  Warning = 'warning',
  Error = 'error',
  Lighten = 'lighten'
}

export enum FontWeight {
  Light = 'light',
  Regular = 'regular',
  Bold = 'bold',
}

export enum Easing {
  EaseIn = 'easeIn',
  EaseOut = 'easeOut',
  EaseInOut = 'easeInOut',
  Sharp = 'sharp',
}

export enum Duration {
  Shortest = 'shortest',
  Shorter = 'shorter',
  Short = 'short',
  Default = 'default',
  Complex = 'complex',
  EnteringScreen = 'enteringScreen',
  LeavingScreen = 'leavingScreen',
}

export enum ElevationLevel {
  None = 'none',
  Primary = 'primary',
  Secondary = 'secondary',
  Modal = 'modal',
}

export enum FlexboxAlignment {
  Baseline = 'baseline',
  Center = 'center',
  FlexEnd = 'flex-end',
  FlexStart = 'flex-start',
  Stretch = 'stretch',
}

export enum FlexboxAlignContent {
  FlexStart = 'flex-start',
  FlexEnd = 'flex-end',
  Center = 'center',
  Stretch = 'stretch',
  SpaceBetween = 'space-between',
  SpaceAround = 'space-around',
}

export enum FlexboxAlignItems {
  FlexStart = 'flex-start',
  FlexEnd = 'flex-end',
  Center = 'center',
  Stretch = 'stretch',
  SpaceBetween = 'space-between',
  SpaceAround = 'space-around',
}

export enum FlexboxDirection {
  Column = 'column',
  ColumnReverse = 'column-reverse',
  Row = 'row',
  RowReverse = 'row-reverse',
}

export enum FlexboxJustification {
  Center = 'center',
  FlexEnd = 'flex-end',
  FlexStart = 'flex-start',
  SpaceAround = 'space-around',
  SpaceBetween = 'space-between',
  SpaceEvenly = 'space-evenly',
}

export enum FlexboxWrap {
  Nowrap = 'nowrap',
  Wrap = 'wrap',
  WrapReverse = 'wrap-reverse',
}

export enum ObjectFit {
  Initial = 'initial',
  Cover = 'cover',
  Contain = 'contain',
  Fill = 'fill',
  ScaleDown = 'scale-down',
  Revert = 'revert',
  Unset = 'unset',
}